scheduler/requirements.txt
Flask==1.1.1
flask-restful==0.3.7
markdown2==2.3.8
pytest==5.0.1
Flask-APScheduler==1.11.0
SQLAlchemy==1.3.6
boto3==1.9.193
awscli==1.16.203
uWSGI==2.0.18
psycopg2==2.8.3